+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 - Add `webpackChunkName` to dynamic imports.
 - Import CSS for bundler.
+- Load Roboto font via CSS by default.
 
 ## 3.1.1 - 2019-09-17